backend/requirements.txt

fastapi==0.128.1 
uvicorn[standard]==0.40.0
SQLAlchemy==2.0.46 
Jinja2==3.1.6
python-multipart==0.0.22 
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
passlib==1.7.4
Pillow==11.0.0